sam4s: watchdog: Simple modification to reproduce a watchdog bug
Bug:
The watchdog interrupt can trigger sometimes with the WDERR status.
This should however not be possible as WDD=WDV, as dictated in
the datasheet (section 17.4: Functional Description).

Reproduce:
Build and flash the adapted watchdog sample. Reboot the device
a couple of times untill the callback print is called. WDT_SR and
WDT_MR will also be logged. The wdg is configured to not trigger
a reset, but instead call a callback, making it easier to spot
the bug.
The shell is available to ease these reboots.

Signed-off-by: Jaro Van Landschoot <jaro.vanlandschoot@basalte.be>